requirements.txt
transformers==4.30.2
fairscale==0.4.13
torch==1.13.0
torchvision==0.14.0
Pillow==9.5.0
opencv-python==4.7.0.72
addict==2.4.0
yapf==0.40.0

